Planner config firestore

+++ ToPickPlacePlanner/Domain/Models/Configuration.cs
@@ -1,7 +1,7 @@
 namespace Domain.Models{
     public class Configuration{
         public int PopulationSize {get;set;}
-        public decimal MutationProbability {get;set;}
+        public double MutationProbability {get;set;}
         public int EliteSize {get;set;}
         public double SolutionPrecision {get;set;}
